// constants/mutation-type.ts
enum MutationType {
    WINDOW_WIDTH = 'windowWidth',
    WINDOW_HEIGHT = 'windowHeight',
    PROJECTS_DATA = 'projectsData',
    ABOUT_DATA = 'aboutData',
    SPLASH_SCREEN_FINISH = 'splashScreenFinish',
    ANIMATION_ENTER = 'animationEnter',
    ALREADY_VISITED = 'alreadyVisited',
    IS_ABOUT_OPEN = 'isAboutOpen',
    IS_UI_OPTIONS_OPEN = 'isUiOptionsOpen',
    SELECTED_FILTER = 'selectedFilter',
    ERROR_PAGE = 'errorPage',
}

export default MutationType